+++ module-apps/application-messages/windows/NewMessage.cpp
@@ -179,7 +179,7 @@ namespace gui
         recipient->setEdges(gui::RectangleEdgeFlags::GUI_RECT_EDGE_NO_EDGES);
         recipient->setInputMode(new InputMode({InputMode::phone}));
         recipient->setFont(style::window::font::mediumbold);
-        recipient->setAlignment(Alignment(gui::Alignment::Horizontal::Left, gui::Alignment::Vertical::Bottom));
+        recipient->setAlignment(Alignment(gui::Alignment::Horizontal::Left, gui::Alignment::Vertical::Center));
         recipient->activatedCallback    = [=](Item &) -> bool { return selectContact(); };
         recipient->focusChangedCallback = [=](Item &) -> bool {
             updateBottomBar();
@@ -215,7 +215,7 @@ namespace gui
         message->setPenFocusWidth(style::window::default_border_focus_w);
         message->setPenWidth(style::window::messages::sms_border_no_focus);
         message->setFont(style::window::font::medium);
-        message->setAlignment(Alignment(gui::Alignment::Horizontal::Left, gui::Alignment::Vertical::Bottom));
+        message->setAlignment(Alignment(gui::Alignment::Horizontal::Left, gui::Alignment::Vertical::Center));
         message->activatedCallback = [=](Item &) -> bool {
             if (!sendSms()) {
                 LOG_ERROR("sendSms failed");